requirements.txt
streamlit==1.33.0
langchain-groq==0.1.0
groq==0.5.0
python-magic-bin==0.4.14
pymupdf==1.24.0
pillow==10.3.0
python-dotenv==1.0.1
httpx==0.27.0